separate test into separate file
- separate tests for homomorphic addition and homomorphic multiplcation
- addition is passing, mult. is not
-
