+++ ring-lwe/src/main.rs
@@ -1,13 +1,16 @@
 mod keygen;
 mod encrypt;
 mod decrypt;
+mod test;
 
-use crate::keygen::{keygen,keygen_string};
-use crate::encrypt::{encrypt,encrypt_string};
-use crate::decrypt::{decrypt,decrypt_string};
+use crate::keygen::keygen_string;
+use crate::encrypt::encrypt_string;
+use crate::decrypt::decrypt_string;
+use crate::test::{test_basic,test_hom_add,test_hom_prod};
 use std::env;
-use ring_lwe::{Parameters,polymul,polyadd,mod_coeffs,polyinv};
+use ring_lwe::Parameters;
 use polynomial_ring::Polynomial;
+
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -35,77 +38,30 @@ fn main() {
             println!("Usage: cargo run -- test <message>");
             return;
         }
-        let message_string = &args[2];
-        let keypair = keygen_string(&params);
-        let pk_string = keypair.get("public").unwrap();
-        let sk_string = keypair.get("secret").unwrap();
-        let ciphertext_string = encrypt_string(&pk_string,message_string,&params);
-        let decrypted_message = decrypt_string(&sk_string,&ciphertext_string,&params);
-        let test_passed = *message_string == decrypted_message;
-        debug_assert!(test_passed, "test failed: {} != {}", *message_string, decrypted_message);
-        println!("test passed");
+        let message = &args[2];
+        test_basic(message,&params);
     }
 
     //test (partially) homomorphic property on two integers
-    if method == "test_hom" {
+    if method == "test_hom_add" {
         if args.len() != 4 && args.len() != 8 {
             println!("Usage: cargo run -- test <message_0> <message_1>");
             return;
         }
-        let (n, q, t, f) = (params.n, params.q, params.t, &params.f);
-        //read the message strings
         let m0_string = &args[2];
         let m1_string = &args[3];
-        //create polynomials from message strings
-        let m0_int: i64 = m0_string.parse().expect("Failed to parse integer.");
-        let m1_int: i64 = m1_string.parse().expect("Failed to parse integer.");
-        let m0_poly = Polynomial::new({
-            let mut v = vec![0i64; n];
-            v[0] = m0_int;
-            v
-        });
-        let m1_poly = Polynomial::new({
-            let mut v = vec![0i64; n];
-            v[0] = m1_int;
-            v
-        });
-        //generate the keypair
-        let keypair = keygen(&params);
-        //get public and secret keys
-        let pk = keypair.0;
-        let sk = keypair.1;
-        //encrypt plaintext messages
-        let u = encrypt(&pk,&m0_poly,&params);
-        let v = encrypt(&pk,&m1_poly,&params);
-        //compute sum of encrypted data
-        let ciphertext_sum = [&u.0 + &v.0, &u.1 + &v.1];
-        //compute product of encrypted data, using non-standard multiplication
-        let c0 = polymul(&v.0,&v.1,q,&f);
-        let u0v1 = &polymul(&u.0,&v.1,q,&f);
-        let u1v0 = &polymul(&u.1,&v.0,q,&f);
-        let c1 = polyinv(&polyadd(u0v1,u1v0,q,&f),q);
-        let c2 = polymul(&u.0,&u.1,q,&f);
-        let c = (c0, c1, c2);
-        //decrypt ciphertext sum u+v
-        let decrypted_sum = decrypt(&sk,&ciphertext_sum,&params);
-        //compute c0 + c1*s + c2*s*s
-        let c1_sk = &polymul(&c.1,&sk,q,&f);
-        let c2_sk_squared = &polymul(&polymul(&c.2,&sk,q,&f),&sk,q,&f);
-        let ciphertext_prod = polyadd(&polyadd(&c.0,c1_sk,q,&f),c2_sk_squared,q,&f);
-        //let delta = q / t, divide coeffs by 1 / delta^2
-        let delta = q / t;
-        let scaled_prod = mod_coeffs(Polynomial::new(ciphertext_prod.coeffs().iter().map(|&coeff| coeff / (delta * delta) ).collect::<Vec<_>>()),q);
-        
-        //print results
-        println!("input polys m1={:?} m2={:?}",m0_poly, m1_poly);
-        println!("m0 len = {:?}",m0_poly);
-        println!("plaintext sum = {}", m0_int + m1_int);
-        println!("decrypted_sum = {}",decrypted_sum);
-        println!("delta = {}",delta);
-        println!("delta^2 = {}",delta * delta);
-        println!("plaintext product = {}", m0_int * m1_int);
-        //println!("ciphertext_prod = {:?}",ciphertext_prod);
-        println!("scaled_product = {:?}",scaled_prod);
+        test_hom_add(m0_string, m1_string, &params);
+    }
+
+    //test (partially) homomorphic property on two integers
+    if method == "test_hom_prod" {
+        if args.len() != 4 && args.len() != 8 {
+            println!("Usage: cargo run -- test <message_0> <message_1>");
+            return;
+        }
+        let m0_string = &args[2];
+        let m1_string = &args[3];
+        test_hom_prod(m0_string, m1_string, &params);
     }
 
     //generate public and secret keys (parameters optional)
